[#142983763] - add save logic for custom lists

+++ KeithLink/Entree.Migrations/Scripts/AppData/List/Tables/ContractHeaders/ContractHeaders_0023_2017.04.00_up.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [List].[ContractHeaders]
 (
-	[Id]				INT			NOT NULL	PRIMARY KEY IDENTITY(1,1),
+	[Id]				BIGINT		NOT NULL	PRIMARY KEY IDENTITY(1,1),
 	[ContractId]		VARCHAR(8)	NULL,
 	[BranchId]			CHAR(3)		NOT NULL,
 	[CustomerNumber]	CHAR(6)		NOT NULL,
