+++ KeithLink/Entree.Migrations/Scripts/AppData/List/Tables/CustomListDetails/CustomListDetails_0031_2017.04.00_up.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [List].[CustomListDetails]
 (
-	[Id]						INT				NOT NULL	PRIMARY KEY IDENTITY(1,1),
-	[ParentCustomListHeaderId]	INT				NOT NULL,
+	[Id]						BIGINT			NOT NULL	PRIMARY KEY IDENTITY(1,1),
+	[ParentCustomListHeaderId]	BIGINT			NOT NULL,
 	[ItemNumber]				CHAR(6)		    NULL,
 	[Each]						BIT				NULL,
     [Par]						DECIMAL (18, 2) NOT NULL	DEFAULT (0),
